tests/wpt/tests/html/canvas/element/layers/2d.layer.invalid-calls.save-beginLayer-restore.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.layer.invalid-calls.save-beginLayer-restore</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.layer.invalid-calls.save-beginLayer-restore</h1>
<p class="desc">Raises exception on save() + beginLayer() + restore().</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
test(t => {
  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  assert_throws_dom("INVALID_STATE_ERR", function() {
    ctx.save();
    ctx.beginLayer();
    ctx.restore();
  });

}, "Raises exception on save() + beginLayer() + restore().");
</script>
